requirements.txt
numpy==1.23.1
Pillow==9.2.0
pydantic==1.9.1
tqdm==4.64.0
typing_extensions==4.3.0
Wave==0.0.2
